Polish file system view and minor fixes
