+++ src/components/containers/modals/DeleteNodeModal.tsx
@@ -12,8 +12,8 @@ interface Props {
 function DeleteNodeModal(props: Props): JSX.Element {
     return (
         <>
-            <div className="z-10 absolute w-full h-full bg-white bg-opacity-40 firefox:bg-opacity-40 backdrop-filter backdrop-blur-sm" onClick={props.closeCallback} />
-            <div className="z-20 absolute left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2">
+            <div className="z-10 fixed left-0 top-0 w-full h-full bg-white bg-opacity-40 firefox:bg-opacity-40 backdrop-filter backdrop-blur-sm" onClick={props.closeCallback} />
+            <div className="z-20 fixed left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2">
                 <div className="rounded-2xl bg-grey-75 blur-md shadow-2xl">
                     <ModalHeader title={`${capitalize(props.node.type)} deletion confirmation`} className="text-center" />
                     <div className="py-8 px-20">
