+++ src/components/navigation/navbar/Navbar.tsx
@@ -1,13 +1,13 @@
 import NavbarItem from "./NavbarItem";
 import {faFile, faFolderOpen, faBell, faUserCircle} from "@fortawesome/free-regular-svg-icons";
-import {faUserShield, faSlidersH, faSignOutAlt} from "@fortawesome/free-solid-svg-icons";
+import {faUserShield, faSignOutAlt} from "@fortawesome/free-solid-svg-icons";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faCloud} from "@fortawesome/free-solid-svg-icons";
 import {Heading3} from "../../text/Headings";
 import Link from "next/link";
 import {useState} from "react";
 import TransferList from "../../transfers/TransferList";
-import {logoutSession} from "../../../util/processes";
+import {terminateSession} from "../../../util/processes";
 import {useRouter} from "next/router";
 
 function Navbar(): JSX.Element {
@@ -36,20 +36,19 @@ function Navbar(): JSX.Element {
                 </Link>
 
                 <NavbarItem name="Files" icon={faFile} activeRoutes={[/^\/files$/]} href="/files" />
-                <NavbarItem name="Tracked&nbsp;folders" activeRoutes={[]} icon={faFolderOpen} />
+                <NavbarItem name="Tracked&nbsp;folders" activeRoutes={[]} icon={faFolderOpen} className="cursor-not-allowed" />
                 <NavbarItem name="Admin&nbsp;panel" activeRoutes={[/^\/users.*$/]} icon={faUserShield} href="/users" />
 
                 <div className="w-full py-6">
                     <hr className="mx-auto w-2/3 text-grey-200" />
                 </div>
 
-                <NavbarItem name="Notifications" activeRoutes={[]} icon={faBell} badge="12" />
-                <NavbarItem name="Settings" activeRoutes={[]} icon={faSlidersH} />
-                <NavbarItem name="My&nbsp;account" activeRoutes={[]} icon={faUserCircle} />
+                <NavbarItem name="Notifications" activeRoutes={[]} icon={faBell} badge="12" className="cursor-not-allowed" />
+                <NavbarItem name="My&nbsp;account" activeRoutes={[]} icon={faUserCircle} className="cursor-not-allowed" />
 
                 <NavbarItem name="Logout" active={false} activeRoutes={[]} icon={faSignOutAlt} className="absolute inset-x-0 bottom-2" action={async () => {
-                    await logoutSession();
-                    await router.push("/");
+                    await terminateSession();
+                    await router.reload();
                 }} />
                 {
                     // FIXME File transfer view
